Renaming for better readability and adding thread safety
